+++ store/gaskv/store_test.go
@@ -73,16 +73,18 @@ func TestGasKVStoreIterator(t *testing.T) {
 	vb := iterator.Value()
 	require.Equal(t, vb, valFmt(2))
 	iterator.Next()
-	require.Equal(t, types.Gas(13377), meter.GasConsumed())
+
+	require.Equal(t, types.Gas(14565), meter.GasConsumed())
 	kc := iterator.Key()
 	require.Equal(t, kc, keyFmt(3))
 	vc := iterator.Value()
 	require.Equal(t, vc, valFmt(0))
 	iterator.Next()
-	require.Equal(t, types.Gas(13446), meter.GasConsumed())
+	require.Equal(t, types.Gas(14667), meter.GasConsumed())
 	require.False(t, iterator.Valid())
 	require.Panics(t, iterator.Next)
-	require.Equal(t, types.Gas(13476), meter.GasConsumed())
+	require.Equal(t, types.Gas(14697), meter.GasConsumed())
+
 	require.NoError(t, iterator.Error())
 
 	reverseIterator := st.ReverseIterator(nil, nil)
@@ -100,7 +102,7 @@ func TestGasKVStoreIterator(t *testing.T) {
 	require.False(t, reverseIterator.Valid())
 	require.Panics(t, reverseIterator.Next)
 
-	require.Equal(t, types.Gas(13782), meter.GasConsumed())
+	require.Equal(t, types.Gas(15135), meter.GasConsumed())
 }
 
 func TestGasKVStoreOutOfGasSet(t *testing.T) {